face-detection/requirements.txt
aiohttp==3.8.1
black==21.9b0
diskcache==5.2.1
cython==0.29.26
h5py==3.4.0
hypercorn==0.13.2
pillow==8.4.0
wandb==0.12.9
lmdb==1.3.0
yapf==0.32.0
natsort==8.0.2
fire==0.4.0
scipy==1.7.3
tqdm==4.61.2
deepspeed==0.5.10
fastapi==0.73.0
flake8==4.0.1
isort==5.9.3
locust==2.4.1
loguru==0.5.3
onnx==1.10.2
opencv-python==4.5.5.62
pytest==6.2.5
pytest-cov==3.0.0
pytest-html==3.1.1
singledispatchmethod==1.0
user-agents==2.2.0
uvicorn==0.15.0
numpy==1.21.2
protobuf==3.16.0 
validators==0.18.2
pandas==1.3.5
uvloop
